Clean up tests
1. move sync tests to its own module.
2. move the code to fire up verbose logging to its own module.
